Separate cluster and policy functions
Also, address use before definition style of
function definitions in files modified.
